+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - #35: Fix PyYAML deprecation warning and YAML loading vulnerability
  - Fixed resolution name (1440p vs 2k)
  - Updated default bitrates
+ - Added definition of 8k resolution
 
 
 ## 0.2.0 (2019-10-14)